Extract abstract spec
